Add test which computes the Hessian matrix and compare it with a stored file
