+++ tests/small/LandIce/AsciiMeshes/Humboldt/CMakeLists.txt
@@ -10,6 +10,8 @@ configure_file(${CMAKE_CURRENT_SOURCE_DIR}/humboldt_2d.exo.4.2
                ${CMAKE_CURRENT_BINARY_DIR}/humboldt_2d.exo.4.2 COPYONLY)
 configure_file(${CMAKE_CURRENT_SOURCE_DIR}/humboldt_2d.exo.4.3
                ${CMAKE_CURRENT_BINARY_DIR}/humboldt_2d.exo.4.3 COPYONLY)
+configure_file(${CMAKE_CURRENT_SOURCE_DIR}/humboldt_contiguous_2d.exo
+               ${CMAKE_CURRENT_BINARY_DIR}/humboldt_contiguous_2d.exo COPYONLY)
 configure_file(${CMAKE_CURRENT_SOURCE_DIR}/basal_friction.ascii
                ${CMAKE_CURRENT_BINARY_DIR}/basal_friction.ascii COPYONLY)
 configure_file(${CMAKE_CURRENT_SOURCE_DIR}/basal_friction_zeroed.ascii
